src/frontend/requirements.txt
openai==1.63.0
streamlit==1.42.0
azure-identity==1.20.0
azure-search-documents==11.5.2
python-dotenv==1.0.1
langchain==0.3.18
langchain-openai==0.3.5
langchain-community==0.3.17
owlready2==0.47
pymupdf==1.25.3